move cube and fact table implementation into data fabric
