+++ cdap-data-fabric/src/main/java/co/cask/cdap/data2/dataset2/lib/cube/DefaultAggregation.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package co.cask.cdap.metrics.store.cube;
+package co.cask.cdap.data2.dataset2.lib.cube;
 
 import co.cask.cdap.api.dataset.lib.cube.CubeFact;
 import com.google.common.collect.ImmutableList;
