+++ cdap-data-fabric/src/test/java/co/cask/cdap/data2/dataset2/lib/cube/DefaultCubeTest.java
@@ -14,13 +14,13 @@
  * the License.
  */
 
-package co.cask.cdap.metrics.store.cube;
+package co.cask.cdap.data2.dataset2.lib.cube;
 
 import co.cask.cdap.api.dataset.lib.cube.Cube;
 import co.cask.cdap.data2.dataset2.lib.table.inmemory.InMemoryMetricsTable;
 import co.cask.cdap.data2.dataset2.lib.table.inmemory.InMemoryTableService;
-import co.cask.cdap.metrics.store.timeseries.EntityTable;
-import co.cask.cdap.metrics.store.timeseries.FactTable;
+import co.cask.cdap.data2.dataset2.lib.timeseries.EntityTable;
+import co.cask.cdap.data2.dataset2.lib.timeseries.FactTable;
 
 import java.util.Collection;
