+++ cdap-watchdog/src/main/java/co/cask/cdap/metrics/store/DefaultMetricStore.java
@@ -33,11 +33,11 @@ import co.cask.cdap.api.metrics.MetricTimeSeries;
 import co.cask.cdap.api.metrics.MetricType;
 import co.cask.cdap.api.metrics.MetricValue;
 import co.cask.cdap.common.conf.Constants;
-import co.cask.cdap.metrics.store.cube.Aggregation;
-import co.cask.cdap.metrics.store.cube.DefaultAggregation;
-import co.cask.cdap.metrics.store.cube.DefaultCube;
-import co.cask.cdap.metrics.store.cube.FactTableSupplier;
-import co.cask.cdap.metrics.store.timeseries.FactTable;
+import co.cask.cdap.data2.dataset2.lib.cube.Aggregation;
+import co.cask.cdap.data2.dataset2.lib.cube.DefaultAggregation;
+import co.cask.cdap.data2.dataset2.lib.cube.DefaultCube;
+import co.cask.cdap.data2.dataset2.lib.cube.FactTableSupplier;
+import co.cask.cdap.data2.dataset2.lib.timeseries.FactTable;
 import com.google.common.base.Supplier;
 import com.google.common.base.Suppliers;
 import com.google.common.collect.ImmutableList;
