+++ cdap-watchdog/src/main/java/co/cask/cdap/metrics/store/upgrade/MetricsEntityCodec.java
@@ -17,7 +17,7 @@
 package co.cask.cdap.metrics.store.upgrade;
 
 import co.cask.cdap.common.utils.ImmutablePair;
-import co.cask.cdap.metrics.store.timeseries.EntityTable;
+import co.cask.cdap.data2.dataset2.lib.timeseries.EntityTable;
 import com.google.common.base.Preconditions;
 import com.google.common.cache.CacheBuilder;
 import com.google.common.cache.CacheLoader;
